pyproject.toml
[project]
name = "flask-api-stock"
version = "0.1.0"
description = "Projeto API RESTful com flask no python 3.11.9"
authors = [
    {name = "Your Name",email = "you@example.com"}
]
readme = "README.md"
requires-python = ">=3.11,<3.12"
dependencies = [
    "flask (>=3.1.2,<4.0.0)",
    "flask-sqlalchemy (>=3.1.1,<4.0.0)",
    "flask-jwt-extended (>=4.7.1,<5.0.0)",
    "black (>=25.12.0,<26.0.0)",
    "flake8 (>=7.3.0,<8.0.0)",
    "isort (>=7.0.0,<8.0.0)",
    "flask-bcrypt (>=1.0.1,<2.0.0)"
]

[tool.poetry]
packages = [{include = "flask_api_stock_python_3"}]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 240

[tool.flake8]
max-line-length = 240
